services/admin/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
boto3==1.34.34
